requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
DiscordTools
agno
sqlalchemy
pgvector
psycopg[binary]
pypdf
google-generativeai
huggingface-hub
sentence-transformers
exa_py